first pass adding in logger instead of printing
